Made var CC less magic

+++ include/http.h
@@ -3,7 +3,7 @@
 #include "parse.h"
 #include <stdint.h>
 
-int http_curl(City_t *city_array, uint8_t *cc);
+int http_curl(City_t *city_array, int short *cc);
 
 
 #endif
