+++ src/http.c
@@ -6,7 +6,7 @@
 #include <stdlib.h>
 #include <stdint.h>
 
-int http_curl(City_t *city_array, uint8_t *cc) {
+int http_curl(City_t *city_array, int short *cc) {
 
     CURL *curl = curl_easy_init();
     if (!curl) {
